requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.5
Django==4.2.7
django-rest-knox==4.2.0
djangorestframework==3.14.0
Pillow==10.1.0
pycparser==2.21
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
